+++ tools/rtorrent/install.sh
@@ -6,10 +6,12 @@ HOME=/home/`whoami`
 # First install rtorrent 
 sudo $CMD install rtorrent
 
-# user config
-ln -s `pwd`/rtorrent.rc  ~/.rtorrent.rc
+# user configs
+if [ ! -L ~/.rtorrent.rc ]; then #软连接文件不存在
+    rm -rf ~/.rtorrent.rc
+    ln -s `pwd`/rtorrent.rc  ~/.rtorrent.rc
+fi
 
-# root config
-sudo ln -s `pwd`/rtorrent.rc  /root/.rtorrent.rc
-
-mkdir ~/Downloads/session
+if [ ! -d ~/Downloads/session ]; then #文件夹不存在
+    mkdir ~/Downloads/session
+fi
