+++ tools/tmux/install.sh
@@ -6,8 +6,15 @@ CMD=dnf
 sudo $CMD install tmux tmuxinator
 
 git clone https://github.com/gpakosz/.tmux.git  tmux
-# user config
-ln -s `pwd`/tmux/.tmux.conf ~/.tmux.conf
-ln -s `pwd`/tmux/.tmux.conf.local ~/.tmux.conf.local
+# user configs
+if [ ! -L ~/.tmux.conf ]; then #软连接文件不存在
+    rm -rf ~/.tmux.conf
+    ln -s `pwd`/tmux/.tmux.conf ~/.tmux.conf
+fi
+
+if [ ! -L ~/.emacs.d ]; then #软连接文件不存在
+    rm -rf ~/.tmux.conf.local
+    ln -s `pwd`/tmux/.tmux.conf.local ~/.tmux.conf.local
+fi
 # root config
 #sudo ln -s `pwd`/tmux.conf /root/.tmux.conf
